+++ pytsite/auth_ui/ep.py
@@ -40,3 +40,32 @@ def profile_edit(args: dict, inp: dict) -> str:
     form = _odm_ui.get_m_form('user', str(profile_owner.id))
 
     return _tpl.render(tpl_name, {'form': form})
+
+
+def follow(args: dict, inp: dict) -> dict:
+    """Follow widget JS endpoint.
+    """
+    op = inp.get('op')
+    uid = inp.get('uid')
+
+    # Does all required arguments present?
+    if not op or not uid:
+        return {'status': False}
+
+    # Is current user authorized
+    current_user = _auth.get_current_user()
+    if current_user.is_anonymous:
+        return {'status': False}
+
+    user = _auth.get_user(uid=inp.get('uid'))
+
+    if op == 'follow':
+        user.f_add('followers', current_user).save()
+        current_user.f_add('follows', user).save()
+        return {'status': True}
+    elif op == 'unfollow':
+        user.f_sub('followers', current_user).save()
+        current_user.f_sub('follows', user).save()
+        return {'status': True}
+    else:
+        return {'status': False}