+++ pytsite/widget/__init__.py
@@ -1,5 +1,9 @@
 """Pytsite Widgets Package Init.
 """
+# Public API
+from . import _button as button, _input as input, _select as select, _static as static
+from ._base import Base
+
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
@@ -14,8 +18,3 @@ def __init():
     lang.register_package(__name__)
 
 __init()
-
-
-# Public API
-from . import _button as button, _input as input, _select as select, _static as static
-from ._base import Base